+++ QuoteCalculator/Models/LoanOffer.cs
@@ -3,7 +3,7 @@
 	public class LoanOffer
 	{
 		public string LenderName { get; set; }
-		public double Rate { get; set; }
-		public double Amount { get; set; }
+		public decimal Rate { get; set; }
+		public decimal Amount { get; set; }
 	}
 }
